requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Diffusion models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# QR code
qrcode[pil]>=7.4.0
pyzbar>=0.1.9

# Cryptography
ecdsa>=0.18.0

# Augmentation
kornia>=0.7.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
